rust: cargo clippy --all-features --fix --allow-no-vcs

+++ rust/src/dns/log.rs
@@ -670,7 +670,7 @@ pub extern "C" fn rs_dns_log_json_answer(tx: &mut DNSTransaction,
     if let &Some(ref response) = &tx.response {
         for query in &response.queries {
             if dns_log_rrtype_enabled(query.rrtype, flags) {
-                return dns_log_json_answer(js, response, flags as u64).is_ok();
+                return dns_log_json_answer(js, response, flags).is_ok();
             }
         }
     }
